KNL-1287 - Regression cannot upload files > 2G
git-svn-id: https://source.sakaiproject.org/svn/kernel/trunk@311557 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ kernel/kernel-impl/src/main/java/org/sakaiproject/content/impl/DbContentService.java
@@ -2453,7 +2453,7 @@ public class DbContentService extends BaseContentService
 				// write the file
 				out = new FileOutputStream(file);
 
-				int byteCount = 0;
+				long byteCount = 0;
 				// chunk
 				byte[] chunk = new byte[STREAM_BUFFER_SIZE];
 				int lenRead;
